Shorter salt, SHA1 and hashed password as binary
I have added support for shorter salt (*4 characters*), support for *SHA1* signing algorthimg with tests and ability to format result of `hash_password/3` as a raw binary.

Longer explanation:

I am creating a new application which needs to be backward-compatible with older system, which uses *PBKDF2*, but with much shorter salt (default is 4), SHA1 and different formatting scheme (it uses *RFC2307* from *LDAP*). In oder to not clutter this library with such arcane scheme I would love to add the necessary parts which allow me to use `comeonin` with `pbkdf2_elixir` and the rest from `hex.pm`.